Move route utils out from feature flag and restructure.
