+++ coral/src/app/features/topics/request/TopicEditRequest.tsx
@@ -29,7 +29,7 @@ import type {
 } from "src/app/features/topics/request/form-schemas/topic-request-form";
 import formSchema from "src/app/features/topics/request/form-schemas/topic-request-form";
 import { generateTopicNameDescription } from "src/app/features/topics/request/utils";
-import { Routes } from "src/app/router_utils";
+import { Routes } from "src/services/router-utils/types";
 import { getAllEnvironmentsForTopicAndAcl } from "src/domain/environment/environment-api";
 import {
   TopicDetailsPerEnv,
